backend/requirements.txt
fastapi
uvicorn
python-multipart
web3
pydantic
python-dotenv
httpx
numpy
requests
solana
solders
authlib
itsdangerous
starlette
authlib
itsdangerous
starlette